+++ src/feagi_mcp/__init__.py
@@ -2,7 +2,7 @@
 
 from typing import Any
 
-__version__ = "0.0.13"
+__version__ = "0.0.14"
 __author__ = "Neuraville Inc."
 __email__ = "feagi@neuraville.com"